+++ edx_django_utils/user/management/commands/manage_user.py
@@ -88,7 +88,7 @@ class Command(BaseCommand):
         user.delete()
 
     @transaction.atomic
-    def handle(  # pylint: disable=arguments-differ
+    def handle(  # pylint: disable=too-many-positional-arguments, arguments-differ
         self, username, email, is_remove, is_staff, is_superuser, groups,
         unusable_password, initial_password_hash, *args, **options
     ):